Experimentally replace epair/jib with netgraph/jngplus(vnet).
My modified version of jng called 'vnet'(badly named) is used.

Bridged configuration (added the optional 5th arg, a gateway address):
  $ sudo bastille create -V vm1 12.1-RELEASE 192.168.1.5 em0 192.168.1.1

Virtual interface/routed configuration is also supported,
where the INTERFACE name is a virtual one and the 5th arg is to specify
an address of the virtual interface (it's a gateway for the jail).
  $ sudo bastille create -V vm2 12.1-RELEASE 10.0.0.5/24 vi0 10.0.0.1